test(encryption): refactor rename connection methods
